+++ super_editor/lib/src/infrastructure/super_textfield/input_method_engine/_ime_text_editing_controller.dart
@@ -36,10 +36,12 @@ class ImeAttributedTextEditingController extends AttributedTextEditingController
     void Function(RawFloatingCursorPoint)? onIOSFloatingCursorChange,
     Brightness keyboardAppearance = Brightness.light,
     TextInputConnectionFactory? inputConnectionFactory,
+    VoidCallback? onImeConnectionChange,
   })  : _realController = controller ?? AttributedTextEditingController(),
         _disposeClientController = disposeClientController,
         _inputConnectionFactory = inputConnectionFactory,
         _onIOSFloatingCursorChange = onIOSFloatingCursorChange,
+        _onImeConnectionChange = onImeConnectionChange,
         _keyboardAppearance = keyboardAppearance {
     _realController.addListener(_onInnerControllerChange);
   }
@@ -90,6 +92,13 @@ class ImeAttributedTextEditingController extends AttributedTextEditingController
     _onIOSFloatingCursorChange = callback;
   }
 
+  VoidCallback? _onImeConnectionChange;
+
+  /// Sets the callback that's invoked whenever the IME conection is opened or closed.
+  set onImeConnectionChange(VoidCallback? callback) {
+    _onImeConnectionChange = callback;
+  }
+
   TextInputConnection? _inputConnection;
   bool _isKeyboardDisplayDesired = false;
 
@@ -124,6 +133,7 @@ class ImeAttributedTextEditingController extends AttributedTextEditingController
     _sendEditingValueToPlatform();
 
     _osCurrentTextEditingValue = _latestTextEditingValueSentToPlatform!;
+    _onImeConnectionChange?.call();
     _log.fine('Is attached to input client? ${_inputConnection!.attached}');
   }
 
@@ -157,6 +167,8 @@ class ImeAttributedTextEditingController extends AttributedTextEditingController
     );
     _inputConnection = _inputConnectionFactory?.call(this, imeConfig) ?? TextInput.attach(this, imeConfig);
     _inputConnection!.show();
+
+    _onImeConnectionChange?.call();
     _sendEditingValueToPlatform();
 
     _osCurrentTextEditingValue = _latestTextEditingValueSentToPlatform!;
@@ -167,6 +179,7 @@ class ImeAttributedTextEditingController extends AttributedTextEditingController
     _inputConnection?.close();
 
     _osCurrentTextEditingValue = const TextEditingValue();
+    _onImeConnectionChange?.call();
   }
 
   void showKeyboard() {
